+++ gqlalchemy/memgraph.py
@@ -15,18 +15,18 @@
 import os
 import sqlite3
 
-from typing import Any, Dict, Iterator, List, Optional, Union
+from typing import List, Optional, Union
 
-from ..connection import Connection, MemgraphConnection
-from ..disk_storage import OnDiskPropertyDatabase
-from ..graph_algorithms.query_modules import QueryModule
-from ..exceptions import (
+from .connection import Connection, MemgraphConnection
+from .database import Database
+from .disk_storage import OnDiskPropertyDatabase
+from .graph_algorithms.query_modules import QueryModule
+from .exceptions import (
     GQLAlchemyError,
     GQLAlchemyUniquenessConstraintError,
     GQLAlchemyOnDiskPropertyDatabaseNotDefinedError,
 )
-from ..models import (
-    MemgraphConstraint,
+from .models import (
     MemgraphConstraintExists,
     MemgraphConstraintUnique,
     MemgraphIndex,
@@ -56,7 +56,7 @@ class MemgraphConstants:
     UNIQUE = "unique"
 
 
-class Memgraph:
+class Memgraph(Database):
     def __init__(
         self,
         host: str = MG_HOST,
@@ -66,34 +66,8 @@ class Memgraph:
         encrypted: bool = MG_ENCRYPTED,
         client_name: str = MG_CLIENT_NAME,
     ):
-        self._host = host
-        self._port = port
-        self._username = username
-        self._password = password
-        self._encrypted = encrypted
-        self._client_name = client_name
-        self._cached_connection: Optional[Connection] = None
+        super().__init__(host, port, username, password, encrypted, client_name)
         self._on_disk_db = None
-
-    def execute_and_fetch(self, query: str, connection: Connection = None) -> Iterator[Dict[str, Any]]:
-        """Executes Cypher query and returns iterator of results."""
-        connection = connection or self._get_cached_connection()
-        return connection.execute_and_fetch(query)
-
-    def execute(self, query: str, connection: Connection = None) -> None:
-        """Executes Cypher query without returning any results."""
-        connection = connection or self._get_cached_connection()
-        connection.execute(query)
-
-    def create_index(self, index: MemgraphIndex) -> None:
-        """Creates an index (label or label-property type) in the database"""
-        query = f"CREATE INDEX ON {index.to_cypher()};"
-        self.execute(query)
-
-    def drop_index(self, index: MemgraphIndex) -> None:
-        """Drops an index (label or label-property type) in the database"""
-        query = f"DROP INDEX ON {index.to_cypher()};"
-        self.execute(query)
 
     def get_indexes(self) -> List[MemgraphIndex]:
         """Returns a list of all database indexes (label and label-property types)"""
@@ -106,29 +80,6 @@ class Memgraph:
                 )
             )
         return indexes
-
-    def ensure_indexes(self, indexes: List[MemgraphIndex]) -> None:
-        """Ensures that database indexes match input indexes"""
-        old_indexes = set(self.get_indexes())
-        new_indexes = set(indexes)
-        for obsolete_index in old_indexes.difference(new_indexes):
-            self.drop_index(obsolete_index)
-        for missing_index in new_indexes.difference(old_indexes):
-            self.create_index(missing_index)
-
-    def drop_indexes(self) -> None:
-        """Drops all indexes in the database"""
-        self.ensure_indexes(indexes=[])
-
-    def create_constraint(self, index: MemgraphConstraint) -> None:
-        """Creates a constraint (label or label-property type) in the database"""
-        query = f"CREATE CONSTRAINT ON {index.to_cypher()};"
-        self.execute(query)
-
-    def drop_constraint(self, index: MemgraphConstraint) -> None:
-        """Drops a constraint (label or label-property type) in the database"""
-        query = f"DROP CONSTRAINT ON {index.to_cypher()};"
-        self.execute(query)
 
     def get_constraints(
         self,
@@ -162,17 +113,17 @@ class Memgraph:
     ) -> List[MemgraphConstraintUnique]:
         return [x for x in self.get_constraints() if isinstance(x, MemgraphConstraintUnique)]
 
-    def ensure_constraints(
-        self,
-        constraints: List[Union[MemgraphConstraintExists, MemgraphConstraintUnique]],
-    ) -> None:
-        """Ensures that database constraints match input constraints"""
-        old_constraints = set(self.get_constraints())
-        new_constraints = set(constraints)
-        for obsolete_constraints in old_constraints.difference(new_constraints):
-            self.drop_constraint(obsolete_constraints)
-        for missing_constraint in new_constraints.difference(old_constraints):
-            self.create_constraint(missing_constraint)
+    def new_connection(self) -> Connection:
+        """Creates new Memgraph connection"""
+        args = dict(
+            host=self._host,
+            port=self._port,
+            username=self._username,
+            password=self._password,
+            encrypted=self._encrypted,
+            client_name=self._client_name,
+        )
+        return MemgraphConnection(**args)
 
     def create_stream(self, stream: MemgraphStream) -> None:
         """Create a stream"""
@@ -195,10 +146,6 @@ class Memgraph:
         """Drop a stream"""
         query = f"DROP STREAM {stream.name};"
         self.execute(query)
-
-    def drop_database(self):
-        """Drops database by removing all nodes and edges"""
-        self.execute("MATCH (n) DETACH DELETE n;")
 
     def create_trigger(self, trigger: MemgraphTrigger) -> None:
         """Creates a trigger"""
@@ -239,25 +186,6 @@ class Memgraph:
         for trigger in self.get_triggers():
             self.drop_trigger(trigger)
 
-    def _get_cached_connection(self) -> Connection:
-        """Returns cached connection if it exists, creates it otherwise"""
-        if self._cached_connection is None or not self._cached_connection.is_active():
-            self._cached_connection = self.new_connection()
-
-        return self._cached_connection
-
-    def new_connection(self) -> Connection:
-        """Creates new Memgraph connection"""
-        args = dict(
-            host=self._host,
-            port=self._port,
-            username=self._username,
-            password=self._password,
-            encrypted=self._encrypted,
-            client_name=self._client_name,
-        )
-        return MemgraphConnection(**args)
-
     def init_disk_storage(self, on_disk_db: OnDiskPropertyDatabase) -> None:
         """Adds and OnDiskPropertyDatabase to Memgraph so that any property
         that has a Field(on_disk=True) can be stored to and loaded from
@@ -268,41 +196,6 @@ class Memgraph:
     def remove_on_disk_storage(self) -> None:
         """Removes the OnDiskPropertyDatabase from Memgraph"""
         self.on_disk_db = None
-
-    def _get_nodes_with_unique_fields(self, node: Node) -> Optional[Node]:
-        """Get's all nodes from Memgraph that have any of the unique fields
-        set to the values in the `node` object.
-        """
-        return self.execute_and_fetch(
-            f"MATCH (node: {node._label})"
-            + f" WHERE {node._get_cypher_unique_fields_or_block('node')}"
-            + " RETURN node;"
-        )
-
-    def get_variable_assume_one(self, query_result: Iterator[Dict[str, Any]], variable_name: str) -> Any:
-        """Returns a single result from the query_result (usually gotten from
-        the execute_and_fetch function).
-        If there is more than one result, raises a GQLAlchemyError.
-        """
-        result = next(query_result, None)
-        next_result = next(query_result, None)
-        if result is None:
-            raise GQLAlchemyError("No result found. Result list is empty.")
-        elif next_result is not None:
-            raise GQLAlchemyError(
-                f"One result expected, but more than one result found. First result: {result}, second result: {next_result}"
-            )
-        elif variable_name not in result:
-            raise GQLAlchemyError(f"Variable name {variable_name} not present in result.")
-
-        return result[variable_name]
-
-    def create_node(self, node: Node) -> Optional[Node]:
-        """Creates a node in Memgraph from the `node` object."""
-        results = self.execute_and_fetch(
-            f"CREATE (node:{node._label}) {node._get_cypher_set_properties('node')} RETURN node;"
-        )
-        return self.get_variable_assume_one(results, "node")
 
     def save_node(self, node: Node) -> Node:
         """Saves node to Memgraph.
@@ -333,11 +226,6 @@ class Memgraph:
         result = self._save_node_properties_on_disk(node, result)
         return result
 
-    def save_nodes(self, nodes: List[Node]) -> None:
-        """Saves a list of nodes to Memgraph."""
-        for i in range(len(nodes)):
-            nodes[i]._id = self.save_node(nodes[i])._id
-
     def _save_node_properties_on_disk(self, node: Node, result: Node) -> Node:
         """Saves all on_disk properties to the on disk database attached to
         Memgraph.
@@ -351,17 +239,6 @@ class Memgraph:
                 setattr(result, field, value)
 
         return result
-
-    def save_node_with_id(self, node: Node) -> Optional[Node]:
-        """Saves a node in Memgraph using the internal Memgraph id."""
-        results = self.execute_and_fetch(
-            f"MATCH (node: {node._label})"
-            + f" WHERE id(node) = {node._id}"
-            + f" {node._get_cypher_set_properties('node')}"
-            + " RETURN node;"
-        )
-
-        return self.get_variable_assume_one(results, "node")
 
     def load_node(self, node: Node) -> Optional[Node]:
         """Loads a node from Memgraph.
@@ -400,19 +277,6 @@ class Memgraph:
                 setattr(result, field, new_value)
 
         return result
-
-    def load_node_with_all_properties(self, node: Node) -> Optional[Node]:
-        """Loads a node from Memgraph with all equal property values."""
-        results = self.execute_and_fetch(
-            f"MATCH (node: {node._label}) WHERE {node._get_cypher_fields_and_block('node')} RETURN node;"
-        )
-        return self.get_variable_assume_one(results, "node")
-
-    def load_node_with_id(self, node: Node) -> Optional[Node]:
-        """Loads a node with the same internal Memgraph id."""
-        results = self.execute_and_fetch(f"MATCH (node: {node._label}) WHERE id(node) = {node._id} RETURN node;")
-
-        return self.get_variable_assume_one(results, "node")
 
     def load_relationship(self, relationship: Relationship) -> Optional[Relationship]:
         """Returns a relationship loaded from Memgraph.
@@ -453,35 +317,6 @@ class Memgraph:
 
         return result
 
-    def load_relationship_with_id(self, relationship: Relationship) -> Optional[Relationship]:
-        """Loads a relationship from Memgraph using the internal id."""
-        results = self.execute_and_fetch(
-            f"MATCH (start_node)-[relationship: {relationship._type}]->(end_node)"
-            + f" WHERE id(start_node) = {relationship._start_node_id}"
-            + f" AND id(end_node) = {relationship._end_node_id}"
-            + f" AND id(relationship) = {relationship._id}"
-            + " RETURN relationship;"
-        )
-        return self.get_variable_assume_one(results, "relationship")
-
-    def load_relationship_with_start_node_id_and_end_node_id(
-        self, relationship: Relationship
-    ) -> Optional[Relationship]:
-        """Loads a relationship from Memgraph using start node and end node id
-        for which all properties of the relationship that are not None match.
-        """
-        and_block = relationship._get_cypher_fields_and_block("relationship")
-        if and_block.strip():
-            and_block = " AND " + and_block
-        results = self.execute_and_fetch(
-            f"MATCH (start_node)-[relationship:{relationship._type}]->(end_node)"
-            + f" WHERE id(start_node) = {relationship._start_node_id}"
-            + f" AND id(end_node) = {relationship._end_node_id}"
-            + and_block
-            + " RETURN relationship;"
-        )
-        return self.get_variable_assume_one(results, "relationship")
-
     def save_relationship(self, relationship: Relationship) -> Optional[Relationship]:
         """Saves a relationship to Memgraph.
         If relationship._id is not None it finds the relationship in Memgraph
@@ -500,11 +335,6 @@ class Memgraph:
         result = self._save_relationship_properties_on_disk(relationship, result)
         return result
 
-    def save_relationships(self, relationships: List[Relationship]) -> None:
-        """Saves a list of relationships to Memgraph."""
-        for i in range(len(relationships)):
-            relationships[i]._id = self.save_relationship(relationships[i])._id
-
     def _save_relationship_properties_on_disk(self, relationship: Relationship, result: Relationship) -> Relationship:
         """Saves on_disk relationship propeties on the OnDiskPropertyDatabase
         added with Memgraph().init_disk_storage(db). If OnDiskPropertyDatabase
@@ -519,32 +349,6 @@ class Memgraph:
                 setattr(result, field, value)
 
         return result
-
-    def save_relationship_with_id(self, relationship: Relationship) -> Optional[Relationship]:
-        """Saves a relationship in Memgraph using the relationship._id."""
-        results = self.execute_and_fetch(
-            f"MATCH (start_node)-[relationship: {relationship._type}]->(end_node)"
-            + f" WHERE id(start_node) = {relationship._start_node_id}"
-            + f" AND id(end_node) = {relationship._end_node_id}"
-            + f" AND id(relationship) = {relationship._id}"
-            + relationship._get_cypher_set_properties("relationship")
-            + " RETURN node;"
-        )
-
-        return self.get_variable_assume_one(results, "relationship")
-
-    def create_relationship(self, relationship: Relationship) -> Optional[Relationship]:
-        """Creates a new relationship in Memgraph."""
-        results = self.execute_and_fetch(
-            "MATCH (start_node), (end_node)"
-            + f" WHERE id(start_node) = {relationship._start_node_id}"
-            + f" AND id(end_node) = {relationship._end_node_id}"
-            + f" CREATE (start_node)-[relationship:{relationship._type}]->(end_node)"
-            + relationship._get_cypher_set_properties("relationship")
-            + "RETURN relationship"
-        )
-
-        return self.get_variable_assume_one(results, "relationship")
 
     def get_procedures(self, starts_with: Optional[str] = None, update: bool = False) -> List["QueryModule"]:
         """Return query procedures.
